// ClientApp/apps/vendor-management/src/Modules/Shared/Interfaces/Request/API-Request-VendorAddress.interface.ts
export interface IAPIRequestVendorAddress {
  userId: string;
  id: number;
  addressTypeId: number;
  cityId: number;
  cityName: string;
  cityCode: string;
  stateId: number;
  stateName: string;
  stateCode: string;
  countryId: number;
  countryCode: string;
  countryName: string;
  zipCode: string;
  addressLine1: string,
  addressLine2: string,
  vendorId: number,
}